ml/ml_local/requirements.txt
fastapi
uvicorn
firebase-admin
sendgrid
pydantic
python-dotenv
